requirements.txt
python-dotenv
discord.py
requests
aiohttp
google-generativeai